Update README.md
This project aims to track and analyze instances where open records (FOIA) laws have been used to request communications and documents from climate scientists, often as a tactic to intimidate or discredit them. Inspired by the work of the Climate Science Legal Defense Fund, the goal is to build a public-facing dataset and accompanying visualizations that illuminate these trends over time, geography, and requester type.

This project was created in the fall of 2024 the context of the introduction to data journalism class at Columbia Graduate School of Journalism. It is a work in progress.